refactor token refresh into helper method
