+++ Api/Operations/SetGear.cs
@@ -35,7 +35,7 @@ namespace Coomes.Equipper.Operations
 
         public async Task Execute(long athleteID, long activityID)
         {
-            var athleteTokens = await GetTokensAndRefreshIfNeeded(athleteID);
+            var athleteTokens = await TokenHelper.GetTokensAndRefreshIfNeeded(_tokenStorage, _tokenProvider, athleteID);
 
             if(await _activityStorage.ContainsResults(athleteID, activityID))
             {
@@ -83,28 +83,6 @@ namespace Coomes.Equipper.Operations
             {
                 _logger.LogWarning(e, "Failed to store activity classifications.");
             }
-        }
-
-        private async Task<AthleteTokens> GetTokensAndRefreshIfNeeded(long athleteID)
-        {
-            var athleteTokens = await _tokenStorage.GetTokens(athleteID);
-
-            if (athleteTokens == null)
-            {
-                throw new SetGearException($"No athlete with ID {athleteID} is registered.");
-            }
-
-            var refreshAt = athleteTokens.ExpiresAtUtc.Subtract(TimeSpan.FromMinutes(5));
-            var now = DateTime.UtcNow;
-            if (refreshAt < now)
-            {
-                var newTokens = await _tokenProvider.RefreshToken(athleteTokens);
-                newTokens.UserID = athleteTokens.UserID;
-                await _tokenStorage.AddOrUpdateTokens(newTokens); // todo: concurrent updates?? take latest expire time?
-                athleteTokens = newTokens;
-            }
-
-            return athleteTokens;
         }
     }
 }
